pyproject.toml
[tool.poetry]
name = "test-flask"
version = "0.1.0"
description = ""
authors = ["omanduj <omanduj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^2.0.1"
requests = "^2.26.0"
PyYAML = "^5.4.1"
nltk = "^3.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
